pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "codebase-graph"
version = "0.1.0"
description = "A live, token-efficient structural map of a codebase for AI coding agents"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "broskees" },
]
dependencies = [
    "cased-kit>=3.5.0",
    "watchfiles>=1.0.0",
    "pathspec>=0.12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
]

[project.scripts]
codebase-graph = "src.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
